Final version

+++ Marcelo-Pedro/simpson.c
@@ -1,6 +1,6 @@
 #include "simpson.h"
 
-double DoubleSimpson (double a, double b, int i, int k, unsigned char *data,  double (*f) (int i, int j, int k, unsigned char *data), double *v){
+double DoubleSimpson (double a, double b, int i, int k, unsigned char *data,  double (*f) (int i, double j, int k, unsigned char *data), double *v){
 	double sac, sab, scb, h = b-a, c = (a+b)/2, err;
 	sab = h/6*(f(i, (int) a, k, data) + 4*f(i, (int) (a+b)/2, k, data) + f(i, (int) b, k, data));
 	sac = (h/2)/6*(f(i, (int) a, k, data) + 4*f(i, (int) (a+c)/2, k, data) + f(i, (int) c, k, data));
@@ -10,7 +10,7 @@ double DoubleSimpson (double a, double b, int i, int k, unsigned char *data,  do
 	return err;
 }
 
-double AdaptiveSimpson (double a, double b, int i, int k, unsigned char *data, double (*f) (int i, int j, int k, unsigned char *data), double tol){
+double AdaptiveSimpson (double a, double b, int i, int k, unsigned char *data, double (*f) (int i, double j, int k, unsigned char *data), double tol){
 	double c = (a + b)/2, err, v;
 	err = DoubleSimpson(a, b, i, k, data, f, &v);
 	if(err < tol)
